RandomSearch hinzugefügt.

+++ CV2_Excersice1/SimilarityMeasure.cpp
@@ -2,34 +2,35 @@
 
 using namespace cv;
 
-long colorSSD(const Mat* firstImage, const Mat* secondImage, Point imagePoint, int windowSize, Point offset)
+double colorSSD(const Mat* firstImage, const Mat* secondImage, Point imagePoint, int windowSize, Point offset)
 {
 	// Return infinity if the image point in the first image is outside the image
 	if (imagePoint.x < 0 || imagePoint.y < 0 ||
 		imagePoint.x >= firstImage->cols || imagePoint.y >= firstImage->rows)
-		return std::numeric_limits<long>::infinity();
+		return std::numeric_limits<double>::infinity();
 	// Return infinity if the image point in the first image is outside the image
 	if (imagePoint.x + offset.x < 0 || imagePoint.y + offset.y < 0 ||
 		imagePoint.x + offset.x >= secondImage->cols || imagePoint.y + offset.y >= secondImage->rows)
-		return std::numeric_limits<long>::infinity();
+		return std::numeric_limits<double>::infinity();
 
-	long ssdPatch = 0;
+	double ssdPatch = 0;
 	// Iterate through the patch
 	for (int cY = -windowSize; cY < windowSize; cY++)
 		for (int cX = -windowSize; cX < windowSize; cX++)
 		{
-			long ssd;
+			double ssd;
 			// Try if all pixel are inside the image
 			try
 			{
+				// TODO: divide the output with the times a pixel was added
 				Point imagePatchPixel = Point(imagePoint.x + cX, imagePoint.y + cY);
 				Point offsetPatchPixel = Point(imagePoint.x + offset.x + cX, imagePoint.y + offset.y + cY);
 
 				Vec3d difference = firstImage->at<Vec3d>(imagePatchPixel.y, imagePatchPixel.x) - 
 								   secondImage->at<Vec3d>(offsetPatchPixel.y, offsetPatchPixel.x);
-				ssd = (long)difference[0] * (long)difference[0] +
-					  (long)difference[1] * (long)difference[1] +
-					  (long)difference[2] * (long)difference[2];
+				ssd = difference[0] * difference[0] +
+					  difference[1] * difference[1] +
+					  difference[2] * difference[2];
 			}
 			catch (Exception e)
 			{
